Remove redundant c-style cast in code (#447)
Remove redundant c-style casts

+++ src/game/shared/entities/items/weapons/CGlock.cpp
@@ -112,7 +112,7 @@ void CGlock::GlockFire(float flSpread, float flCycleTime, bool fUseAutoAim)
 
 	m_iClip--;
 
-	m_pPlayer->pev->effects = (int)(m_pPlayer->pev->effects) | EF_MUZZLEFLASH;
+	m_pPlayer->pev->effects = m_pPlayer->pev->effects | EF_MUZZLEFLASH;
 
 	int flags;
