+++ src/game/shared/entities/items/weapons/CMP5.cpp
@@ -130,8 +130,7 @@ void CMP5::PrimaryAttack()
 
 	m_iClip--;
 
-
-	m_pPlayer->pev->effects = (int)(m_pPlayer->pev->effects) | EF_MUZZLEFLASH;
+	m_pPlayer->pev->effects = m_pPlayer->pev->effects | EF_MUZZLEFLASH;
 
 	// player "shoot" animation
 	m_pPlayer->SetAnimation(PLAYER_ATTACK1);
